[AM][OPS-9321] move Income and IncomeCategory to ssttpaffordability.model
