+++ app/ssttpaffordability/model/Income.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ssttpaffordability.model
+
+import play.api.libs.json.{Json, OFormat}
+
+final case class Income(categories: Seq[IncomeCategory]) {
+  def totalIncome: BigDecimal = categories.map(_.amount).sum
+
+  def amount(category: String): BigDecimal = {
+    categories.find(_.category == category).fold(BigDecimal(0))(_.amount)
+  }
+}
+
+object Income {
+  implicit val format: OFormat[Income] = Json.format[Income]
+}
